src/pyproject.toml
[project]
name = "rss-notice-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.5",
    "feedparser>=6.0.12",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
]
